requirements.txt
requests>=2.31.0
websockets>=11.0.3
pandas>=2.0.3
numpy>=1.25.1
python-dotenv>=1.0.0
eth-account>=0.8.0
web3>=6.4.0
openai>=1.0.0
